change point virgule

+++ controllers/FactureFicheNavette.js
@@ -129,7 +129,7 @@ exports.createfacture = async (req, res) => {
       await pool
         .request()
         .input("codechantier", getSql().VarChar, req.body.codechantier)
-        .input("montantAvance", getSql().Float(8,2), req.body.montantAvance)
+        .input("montantAvance",getSql().Numeric(10, 2), req.body.montantAvance)
 
         .input("idfournisseur", getSql().Int, req.body.idfournisseur)
         .input("idFacture", getSql().Int, req.body.idFacture)
